requirements.txt
# 数据库连接
psycopg2-binary>=2.9.7

# 环境变量配置
python-dotenv==1.0.0

# 进度条显示
tqdm>=4.66.1

# 类型检查（开发依赖）
mypy>=1.0.0

# 代码格式化（开发依赖）
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
